refactor(tests): split output rows suite by feature
Replace the monolithic output rows test module with focused core, mail, and service row suites while preserving existing row-builder coverage and helper assertions.
